web/requirements.txt
bidict==0.23.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
gunicorn==22.0.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
phonenumbers==8.13.31
pillow==10.2.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-engineio==4.9.0
python-socketio==5.11.1
qrcode==7.4.2
simple-websocket==1.0.0
typing_extensions==4.10.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
